LayoutTests/inspector/canvas/console-record-offscreen-webgl2-expected.txt

Test that CanvasManager is able to record actions made to offscreen WebGL2 canvas contexts.


== Running test suite: Canvas.recordingOffscreenWebGL2
-- Running test case: Canvas.recordingOffscreenWebGL2.Console
PASS: The recording should have the name "TEST".
PASS: The recording should have one frame.
PASS: The first frame should have one action.
